Move metainfo config
