+++ internal/testplanet/uplink.go
@@ -6,25 +6,19 @@ package testplanet
 import (
 	"bytes"
 	"context"
-	"fmt"
 	"io"
 	"io/ioutil"
 
-	"github.com/vivint/infectious"
 	"github.com/zeebo/errs"
 	"go.uber.org/zap"
 	"google.golang.org/grpc"
 
 	"storj.io/storj/internal/memory"
-	"storj.io/storj/pkg/eestream"
 	"storj.io/storj/pkg/identity"
-	"storj.io/storj/pkg/metainfo/kvmetainfo"
+	"storj.io/storj/pkg/metainfo"
 	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/pointerdb/pdbclient"
-	"storj.io/storj/pkg/storage/buckets"
-	ecclient "storj.io/storj/pkg/storage/ec"
-	"storj.io/storj/pkg/storage/segments"
 	"storj.io/storj/pkg/storage/streams"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/pkg/stream"
@@ -113,13 +107,14 @@ func (uplink *Uplink) DialOverlay(destination Peer) (overlay.Client, error) {
 
 // Upload data to specific satellite
 func (uplink *Uplink) Upload(ctx context.Context, satellite *satellite.Peer, bucket string, path storj.Path, data []byte) error {
-	metainfo, streams, err := uplink.getMetainfo(satellite)
+	config := uplink.getDefaultConfig(satellite)
+	metainfo, streams, err := config.GetMetainfo(ctx, uplink.Identity)
 	if err != nil {
 		return err
 	}
 
-	encScheme := uplink.getEncryptionScheme()
-	redScheme := uplink.getRedundancyScheme()
+	encScheme := config.GetEncryptionScheme()
+	redScheme := config.GetRedundancyScheme()
 
 	// create bucket if not exists
 	_, err = metainfo.GetBucket(ctx, bucket)
@@ -167,7 +162,8 @@ func uploadStream(ctx context.Context, streams streams.Store, mutableObject stor
 
 // Download data from specific satellite
 func (uplink *Uplink) Download(ctx context.Context, satellite *satellite.Peer, bucket string, path storj.Path) ([]byte, error) {
-	metainfo, streams, err := uplink.getMetainfo(satellite)
+	config := uplink.getDefaultConfig(satellite)
+	metainfo, streams, err := config.GetMetainfo(ctx, uplink.Identity)
 	if err != nil {
 		return []byte{}, err
 	}
@@ -187,75 +183,26 @@ func (uplink *Uplink) Download(ctx context.Context, satellite *satellite.Peer, b
 	return data, nil
 }
 
-func (uplink *Uplink) getMetainfo(satellite *satellite.Peer) (db storj.Metainfo, ss streams.Store, err error) {
-	encScheme := uplink.getEncryptionScheme()
-	redScheme := uplink.getRedundancyScheme()
+func (uplink *Uplink) getDefaultConfig(satellite *satellite.Peer) metainfo.Config {
+	return metainfo.Config{
+		Enc: metainfo.EncryptionConfig{
+			DataType:  int(storj.AESGCM),
+			BlockSize: 1 * memory.KiB,
+		},
+		RS: metainfo.RSConfig{
+			MinThreshold:     1 * uplink.StorageNodeCount / 5,
+			RepairThreshold:  2 * uplink.StorageNodeCount / 5,
+			SuccessThreshold: 3 * uplink.StorageNodeCount / 5,
+			MaxThreshold:     4 * uplink.StorageNodeCount / 5,
 
-	// redundancy settings
-	minThreshold := int(redScheme.RequiredShares)
-	repairThreshold := int(redScheme.RepairShares)
-	successThreshold := int(redScheme.OptimalShares)
-	maxThreshold := int(redScheme.TotalShares)
-	erasureShareSize := 1 * memory.KiB
-	maxBufferMem := 4 * memory.MiB
-
-	// client settings
-	maxInlineSize := 4 * memory.KiB
-	segmentSize := 64 * memory.MiB
-
-	oc, err := uplink.DialOverlay(satellite)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	pdb, err := uplink.DialPointerDB(satellite, "") // TODO pass api key?
-	if err != nil {
-		return nil, nil, err
-	}
-
-	ec := ecclient.NewClient(uplink.Identity, maxBufferMem.Int())
-	fc, err := infectious.NewFEC(minThreshold, maxThreshold)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	rs, err := eestream.NewRedundancyStrategy(eestream.NewRSScheme(fc, erasureShareSize.Int()), repairThreshold, successThreshold)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	segments := segments.NewSegmentStore(oc, ec, pdb, rs, maxInlineSize.Int())
-
-	if erasureShareSize.Int()*minThreshold%int(encScheme.BlockSize) != 0 {
-		return nil, nil, fmt.Errorf("EncryptionBlockSize must be a multiple of ErasureShareSize * RS MinThreshold")
-	}
-
-	key := new(storj.Key)
-	copy(key[:], "enc.key")
-
-	streams, err := streams.NewStreamStore(segments, segmentSize.Int64(), key, int(encScheme.BlockSize), encScheme.Cipher)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	buckets := buckets.NewStore(streams)
-
-	return kvmetainfo.New(buckets, streams, segments, pdb, key), streams, nil
-}
-
-func (uplink *Uplink) getRedundancyScheme() storj.RedundancyScheme {
-	return storj.RedundancyScheme{
-		Algorithm:      storj.ReedSolomon,
-		RequiredShares: int16(1 * uplink.StorageNodeCount / 5),
-		RepairShares:   int16(2 * uplink.StorageNodeCount / 5),
-		OptimalShares:  int16(3 * uplink.StorageNodeCount / 5),
-		TotalShares:    int16(4 * uplink.StorageNodeCount / 5),
-	}
-}
-
-func (uplink *Uplink) getEncryptionScheme() storj.EncryptionScheme {
-	return storj.EncryptionScheme{
-		Cipher:    storj.AESGCM,
-		BlockSize: 1 * memory.KiB.Int32(),
+			ErasureShareSize: 1 * memory.KiB,
+			MaxBufferMem:     4 * memory.MiB,
+		},
+		Client: metainfo.ClientConfig{
+			OverlayAddr:   satellite.Addr(),
+			PointerDBAddr: satellite.Addr(),
+			MaxInlineSize: 4 * memory.KiB,
+			SegmentSize:   64 * memory.MiB,
+		},
 	}
 }
